--- assignment/src/assignment2/PersonDetails.java
@@ -1,12 +1,0 @@
-package assignment2;
-
-public abstract class PersonDetails {
-    abstract String getName();
-    abstract void setName(String name);
-    abstract String getAddress();
-    abstract void setAddress(String address);
-    abstract String getRole();
-    abstract void setRole(String role);
-    abstract String getResponsibilities();
-    abstract void setResponsibilities(String responsibilities);
-}
