--- assignment/src/assignment3/BikeSpeed.java
@@ -1,8 +1,0 @@
-package assignment3;
-
-public interface BikeSpeed {
-    //using final keyword to restrict bike speed
-    final int speed = 90;
-    int getSpeed();
-    void manufacturingFigures();
-}
