--- assignment/src/com/assignment/Assignment.java
@@ -1,7 +1,0 @@
-package com.assignment;
-
-public class Assignment {
-    public static void main(String[] args) {
-        System.out.println("Hello World by Siva");
-    }
-}
