--- assignment/src/com/assignment/BlackAndWhiteTV.java
@@ -1,7 +1,0 @@
-package com.assignment;
-
-public class BlackAndWhiteTV {
-    public void selection(){
-        System.out.println("You have selected Black and White TV");
-    }
-}
