--- assignment/src/com/assignment/ColourTV.java
@@ -1,9 +1,0 @@
-package com.assignment;
-
-import com.assignment.BlackAndWhiteTV;
-
-public class ColourTV extends BlackAndWhiteTV {
-    public void selction() {
-        System.out.println("You have selected colour TV");
-    }
-}
