+++ assignment/src/com/assignment/week1/Assignment.java
@@ -1,0 +1,7 @@
+package com.assignment.week1;
+
+public class Assignment {
+    public static void main(String[] args) {
+        System.out.println("Hello World by Siva");
+    }
+}
