+++ assignment/src/com/assignment/week1/BlackAndWhiteTV.java
@@ -1,0 +1,7 @@
+package com.assignment.week1;
+
+public class BlackAndWhiteTV {
+    public void selection(){
+        System.out.println("You have selected Black and White TV");
+    }
+}
