+++ assignment/src/com/assignment/week1/ColourTV.java
@@ -1,0 +1,7 @@
+package com.assignment.week1;
+
+public class ColourTV extends BlackAndWhiteTV {
+    public void selction() {
+        System.out.println("You have selected colour TV");
+    }
+}
