+++ assignment/src/com/assignment/week2/abstractClass/PersonDetails.java
@@ -1,0 +1,12 @@
+package com.assignment.week2.abstractClass;
+
+public abstract class PersonDetails {
+    abstract String getName();
+    abstract void setName(String name);
+    abstract String getAddress();
+    abstract void setAddress(String address);
+    abstract String getRole();
+    abstract void setRole(String role);
+    abstract String getResponsibilities();
+    abstract void setResponsibilities(String responsibilities);
+}
