+++ assignment/src/com/assignment/week2/interfaceConcept/BikeSpeed.java
@@ -1,0 +1,8 @@
+package com.assignment.week2.interfaceConcept;
+
+public interface BikeSpeed {
+    //using final keyword to restrict bike speed
+    final int speed = 90;
+    int getSpeed();
+    void manufacturingFigures();
+}
