+++ assignment/src/com/assignment/week2/interfaceConcept/Hero.java
@@ -1,0 +1,15 @@
+package com.assignment.week2.interfaceConcept;
+
+public class Hero implements BikeSpeed {
+    int speed = 60;
+    //implementing methods from the interface
+    @Override
+    public int getSpeed() {
+        return BikeSpeed.speed;
+    }
+
+    @Override
+    public void manufacturingFigures() {
+        System.out.println("50000 bikes are manufactured this month");
+    }
+}
