+++ assignment/src/com/assignment/week2/interfaceConcept/Honda.java
@@ -1,0 +1,16 @@
+package com.assignment.week2.interfaceConcept;
+
+public class Honda implements BikeSpeed {
+    int speed = 60;
+    //implementing methods from the interface
+    @Override
+    public int getSpeed() {
+        return BikeSpeed.speed;
+    }
+
+    @Override
+    public void manufacturingFigures() {
+        System.out.println("30000 bikes are manufactured this month");
+
+    }
+}
