+++ assignment/src/com/assignment/week3/exceptions2/ArrayIndexOutOfBoundException.java
@@ -1,0 +1,19 @@
+package com.assignment.week3.exceptions2;
+
+public class ArrayIndexOutOfBoundException {
+    public static void main(String[] args) {
+
+        int arr[] = new int[]{1, 5,9,7,3};
+
+        System.out.println("Index is 2 = "+arr[2]);
+
+        //catching ArrayIndexOutOfBoundsException exception in catch block
+        try {
+            System.out.println("Index is 6 = " + arr[6]);
+        }
+        catch(ArrayIndexOutOfBoundsException e) {
+            System.out.println("End of main");
+        }
+    }
+
+}
