+++ assignment/src/com/assignment/week3/exceptions3/target_file.txt
@@ -1,0 +1,3 @@
+Writing the content to be transferred to the target file...
+Hi!!! This is the source file's text...
+Now it is parsed to target file successfully
