+++ assignment/src/com/assignment/week3/lambdaexpressions1/Employee.java
@@ -1,0 +1,16 @@
+package com.assignment.week3.lambdaexpressions1;
+
+public class Employee {
+
+    //Declaring required fields
+    public int employeeId;
+    public String department;
+    public int salary;
+
+    //Creating a constructor that instantiates this class
+    public Employee (int employeeId, String department, int salary) {
+        this.employeeId = employeeId;
+        this.department = department;
+        this.salary = salary;
+    }
+}
