+++ assignment/src/com/assignment/week3/lambdaexpressions2/iSquaringNumber.java
@@ -1,0 +1,9 @@
+package com.assignment.week3.lambdaexpressions2;
+
+//The interface with only one method is known as Functional Interface
+@FunctionalInterface
+public interface iSquaringNumber {
+    //declaring the method to be implemented
+    int squares(int number);
+}
+
