+++ assignment/src/com/assignment/week3/thread2/BankAccount.java
@@ -1,0 +1,17 @@
+package com.assignment.week3.thread2;
+
+public class BankAccount {
+    private double balance;
+
+    //Withdraws amount from the bank account
+    public double withdraw(double amount) {
+        balance -= amount;
+        return balance;
+    }
+
+    //Deposits amount to the bank account
+    public double deposit(double amount) {
+        balance += amount;
+        return balance;
+    }
+}
